Add fingerprint to gcp_compute_url_map
This commit adds the `fingerprint` parameter
to the `gcp_compute_url_map` module so that
updates to url maps can be processed.

Without this commit, we'll run into something like
the following when trying to update url maps.

```
Required field 'resource.fingerprint' not specified
```

This is a quick and dirty hack as this change should
really be made (honestly may already be fixed in newer
versions) in the magic modules repo as the file
changed in this commit is generated from an API spec.

That said, this allows the url map module to function
mostly as expected for the time being.

Either way, major changes appear to be coming to the
collections, as well as the general gcp load balancing
modules, so hopefully this is short lived.

Signed-off-by: Evan Gray <evanscottgray@gmail.com>